+++ .gitignore
@@ -1,11 +1,3 @@
-.DS_Store
-.dart_tool/
-
-.packages
-.pub/
-
-build/
-
 # Miscellaneous
 *.class
 *.lock
