+++ android/build.gradle
@@ -1,4 +1,4 @@
-group 'com.chabokpush.flutter.chabokpush'
+group 'io.chabokpush.flutter'
 version '1.0'
 
 buildscript {
@@ -35,7 +35,7 @@ android {
 
 dependencies {
     // api fileTree(include: ['*.jar', '*.aar'], dir: 'libs')
-    api 'com.adpdigital.push:chabok-lib:3.4.1'
+    api 'com.adpdigital.push:chabok-lib:3.5.0'
 
     implementation 'me.leolin:ShortcutBadger:1.1.22@aar'
     implementation 'com.google.firebase:firebase-messaging:17.1.0'
