+++ android/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.chabokpush.flutter.chabokpush">
+    package="io.chabokpush.flutter">
 
     <application>
         <meta-data
