+++ example/lib/main.dart
@@ -2,9 +2,9 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 
-import 'package:chabokpush/chabokpush.dart';
-import 'package:chabokpush/ChabokEvent.dart';
-import 'package:chabokpush/ChabokMessage.dart';
+import 'package:chabokpush_flutter/chabokpush.dart';
+import 'package:chabokpush_flutter/ChabokEvent.dart';
+import 'package:chabokpush_flutter/ChabokMessage.dart';
 
 void main() => runApp(MyApp());
