+++ ios/chabokpush_flutter.podspec
@@ -1,10 +1,10 @@
 #
 # To learn more about a Podspec see http://guides.cocoapods.org/syntax/podspec.html.
-# Run `pod lib lint chabokpush.podspec' to validate before publishing.
+# Run pod lib lint 'chabokpush.podspec' to validate before publishing.
 #
 Pod::Spec.new do |s|
-  s.name             = 'chabokpush'
-  s.version          = '1.0.3'
+  s.name             = 'chabokpush_flutter'
+  s.version          = '2.0.0'
   s.summary          = 'Chabok flutter plugin.'
   s.description      = <<-DESC
 Chabok provides your app with in-app messaging and easy geo-location features.
@@ -12,10 +12,7 @@ Chabok provides your app with in-app messaging and easy geo-location features.
 Chabok co
                        DESC
   s.homepage         = 'http://chabok.io'
-  s.license          = { :type => 'proprietary', :text => <<-LICENSE
-Copyright 2020 - present Chabok. All rights reserved.
-LICENSE
-						}	
+  s.license          = { :file => '../LICENSE' }
   s.author           = { 'Chabok Realtime Solutions' => 'info@chabok.io' }
   s.source           = { :path => '.' }
 
@@ -23,7 +20,7 @@ LICENSE
   s.public_header_files = 'Classes/**/*.h'
 
   s.dependency 'Flutter'
-  s.dependency "ChabokPush", "~> 2.2.1"
+  s.dependency "ChabokPush", "~> 2.3.0"
 
   s.platform = :ios, '8.0'
   s.static_framework = true
