+++ pubspec.yaml
@@ -1,6 +1,6 @@
-name: chabokpush
+name: chabokpush_flutter
 description: Chabok Push provides your app with in-app messaging and easy geo-location features.
-version: 1.0.4
+version: 2.0.0
 homepage: https://github.com/chabok-io/chabok-client-flutter
 
 environment:
@@ -27,7 +27,7 @@ flutter:
   plugin:
     platforms:
       android:
-        package: com.chabokpush.flutter.chabokpush
+        package: io.chabokpush.flutter
         pluginClass: ChabokpushPlugin
       ios:
         pluginClass: ChabokpushPlugin
